requirements.txt
python-telegram-bot 8551641638:AAErpiBXorQKKYV2xjhoF9hQr2h6rIHz6Sk
open ai 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